refactor: update structure folder management and rename with snake_case when importing.
