+++ src/App.jsx
@@ -1,25 +1,24 @@
 import { Route, Routes } from "react-router-dom";
-import Login from "@/pages/LoginPage";
-import AdminLayout from "./layouts/AdminLayout";
-import Admin from "./pages/Admin";
-import ProductsManagement from "./pages/Admin/shared/ProductsManagement";
-import CategoriesManagement from "./pages/Admin/shared/CategoriesManagement";
-import OrdersManagement from "./pages/Admin/shared/OrdersManagement";
-import UsersManagement from "./pages/Admin/shared/UsersManagement";
-import BranchesManagement from "./pages/Admin/shared/BranchesManagement";
-import PromotionsManagement from "./pages/Admin/shared/PromotionsManagement";
-import PaymentHistoriesManagement from "./pages/Admin/shared/PaymentHistoriesManagement";
-import RolesManagement from "./pages/Admin/shared/RolesManagement";
-import PermissionsManagement from "./pages/Admin/shared/PermissionsManagement";
-import ColorsManagement from "./pages/Admin/shared/ColorsManagement";
-import WardsManagement from "./pages/Admin/shared/WardsManagement";
-import DistrictsManagement from "./pages/Admin/shared/DistrictsManagement";
-import ProvincesManagement from "./pages/Admin/shared/ProvincesManagement";
-import ProductImagesManagement from "./pages/Admin/shared/ProductImagesManagement";
-import ProductVariantsManagement from "./pages/Admin/shared/ProductVariantsManagement";
-import AddressManagement from "./pages/Admin/shared/AddressManagement";
-import { PrivateRoute, PublicRoute } from "./components/ProtectedRoute";
-import PageNotFound from "./pages/Admin/PageNotFound";
+import Login from "@/pages/Login_page";
+import { PrivateRoute, PublicRoute } from "./components/Protected_route";
+
+import AdminPage from "./pages/Admin_page";
+import AdminLayout from "./layouts/Admin_layout";
+import CategoryManagement from "./pages/Admin_page/shared/Category_management";
+import BranchesManagement from "./pages/Admin_page/shared/Branch_management";
+import PromotionsManagement from "./pages/Admin_page/shared/Promotion_management";
+import PaymentHistoriesManagement from "./pages/Admin_page/shared/Payment_history_management";
+import RolesManagement from "./pages/Admin_page/shared/Role_management";
+import ColorsManagement from "./pages/Admin_page/shared/Color_management";
+import AddressManagement from "./pages/Admin_page/shared/Address_management";
+import PageNotFound from "./pages/Admin_page/Page_not_found";
+import FileManagement from "./pages/Admin_page/shared/File_Management";
+import ProductManagement from "./pages/Admin_page/shared/Product_management";
+import ColorManagement from "./pages/Admin_page/shared/Color_management";
+import BranchManagement from "./pages/Admin_page/shared/Branch_management";
+import OrderManagement from "./pages/Admin_page/shared/Order_management";
+import ProductVariantManagement from "./pages/Admin_page/shared/Product_variant_management";
+import UserManagement from "./pages/Admin_page/shared/User_management";
 
 const App = () => {
   return (
@@ -36,39 +35,26 @@ const App = () => {
       {/* TODO: Set PrivateRoute if the browser not found token or expired token it will return the page login */}
       <Route element={<PrivateRoute />}>
         <Route path="/admin" element={<AdminLayout />}>
-          <Route path="dashboard" element={<Admin />} />
-          <Route path="productsManagement" element={<ProductsManagement />} />
+          <Route path="dashboard-management" element={<AdminPage />} />
+          <Route path="product-management" element={<ProductManagement />} />
+          <Route path="file-management" element={<FileManagement />} />
           <Route
-            path="productImagesManagement"
-            element={<ProductImagesManagement />}
+            path="product-variant-management"
+            element={<ProductVariantManagement />}
           />
+          <Route path="color-management" element={<ColorManagement />} />
+          <Route path="address-management" element={<AddressManagement />} />
+          <Route path="category-management" element={<CategoryManagement />} />
+          <Route path="order-management" element={<OrderManagement />} />
+          <Route path="user-management" element={<UserManagement />} />
+          <Route path="branch-management" element={<BranchManagement />} />
           <Route
-            path="productVariantsManagement"
-            element={<ProductVariantsManagement />}
-          />
-          <Route path="colorsManagement" element={<ColorsManagement />} />
-          <Route path="wardsManagement" element={<WardsManagement />} />
-          <Route path="districtsManagement" element={<DistrictsManagement />} />
-          <Route path="provincesManagement" element={<ProvincesManagement />} />
-          <Route path="addressManagement" element={<AddressManagement />} />
-          <Route
-            path="categoriesManagement"
-            element={<CategoriesManagement />}
-          />
-          <Route path="ordersManagement" element={<OrdersManagement />} />
-          <Route path="usersManagement" element={<UsersManagement />} />
-          <Route path="branchesManagement" element={<BranchesManagement />} />
-          <Route
-            path="promotionsManagement"
+            path="promotion-management"
             element={<PromotionsManagement />}
           />
-          <Route path="rolesManagement" element={<RolesManagement />} />
+          <Route path="role-management" element={<RolesManagement />} />
           <Route
-            path="permissionsManagement"
-            element={<PermissionsManagement />}
-          />
-          <Route
-            path="paymentHistoriesManagement"
+            path="payment-history-management"
             element={<PaymentHistoriesManagement />}
           />
         </Route>
