--- src/pages/Admin/shared/ColorsManagement/index.jsx
@@ -1,309 +1,0 @@
-import { useState } from "react";
-import {
-  Typography,
-  Button,
-  TextField,
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions,
-  IconButton,
-  Box,
-} from "@mui/material";
-import DashboardLayoutWrapper from "@/layouts/DashboardLayout";
-import {
-  useGetAllColorsQuery,
-  useAddColorMutation,
-  useUpdateColorMutation,
-  useDeleteColorMutation,
-} from "@/services/api/color";
-import { Add, Delete, Edit, Refresh } from "@mui/icons-material";
-import { useSnackbar } from "@/components/Snackbar";
-import TableData from "@/components/TableData";
-
-const ColorsManagement = () => {
-  const { showSnackbar } = useSnackbar();
-  const [openDeleteDialog, setOpenDeleteDialog] = useState(false);
-  const [openAddDialog, setOpenAddDialog] = useState(false);
-  const [openUpdateDialog, setOpenUpdateDialog] = useState(false);
-  const [selectedColorId, setSelectedColorId] = useState(null);
-  const [submitted, setSubmitted] = useState(false);
-  const [paginationModel, setPaginationModel] = useState({
-    page: 0,
-    pageSize: 10,
-  });
-  const [newColor, setNewColor] = useState({ name: "" });
-
-  const {
-    data: dataColor,
-    isLoading: isLoadingColor,
-    isError: isErrorColor,
-    error: errorColor,
-    refetch: refetchColor,
-  } = useGetAllColorsQuery(
-    {
-      page: paginationModel.page + 0,
-      size: paginationModel.pageSize,
-    },
-    {
-      refetchOnMountOrArgChange: true,
-    },
-  );
-  const [addColor] = useAddColorMutation();
-  const [updateColor] = useUpdateColorMutation();
-  const [deleteColor] = useDeleteColorMutation();
-
-  const dataRowColors = dataColor?.result?.items || [];
-  const totalRows = dataColor?.result?.totalItems || 0;
-
-  const columnsColor = [
-    { field: "id", headerName: "ID", width: 100 },
-    { field: "name", headerName: "Tên màu sắc", width: 200 },
-    {
-      field: "actions",
-      headerName: "Hành động",
-      width: 200,
-      renderCell: (params) => (
-        <>
-          <IconButton onClick={() => handleEditColor(params.row.id)}>
-            <Edit color="primary" />
-          </IconButton>
-          <IconButton
-            onClick={() => handleOpenDeleteDialog(params.row.id)}
-            color="error"
-          >
-            <Delete />
-          </IconButton>
-        </>
-      ),
-    },
-  ];
-
-  const handleOpenDeleteDialog = (id) => {
-    setSelectedColorId(id);
-    setOpenDeleteDialog(true);
-  };
-
-  const handleCloseDeleteDialog = () => {
-    setSelectedColorId(null);
-    setOpenDeleteDialog(false);
-  };
-
-  const handleRefresh = () => {
-    refetchColor();
-    showSnackbar("Danh sách màu sắc đã được làm mới!", "info");
-  };
-
-  const handleAddColor = async (data) => {
-    setSubmitted(true);
-
-    try {
-      await addColor({
-        name: data?.name,
-      }).unwrap();
-      showSnackbar("Thêm màu sắc thành công!", "success");
-      setNewColor({ name: "" });
-      setOpenAddDialog(false);
-      setSubmitted(false);
-      refetchColor();
-    } catch (error) {
-      if (error && error.data && error.data.message) {
-        showSnackbar(error.data.message, "error");
-      }
-    }
-  };
-
-  const handleEditColor = (id) => {
-    const colorsEdit = dataRowColors.find((item) => item.id === id);
-    console.log("colorEdit", colorsEdit);
-
-    if (colorsEdit) {
-      setNewColor({
-        name: colorsEdit.name,
-      });
-    }
-    setSelectedColorId(id);
-    setOpenUpdateDialog(true);
-  };
-
-  const handleUpdateColor = async () => {
-    setSubmitted(true);
-
-    try {
-      console.log("Selected Color ID:", selectedColorId);
-      await updateColor({
-        id: selectedColorId,
-        ...newColor,
-      }).unwrap();
-      showSnackbar("Cập nhật màu sắc thành công!", "success");
-      setNewColor({ name: "" });
-      setOpenUpdateDialog(false);
-      setSubmitted(false);
-      setSelectedColorId(null);
-      refetchColor();
-    } catch (error) {
-      if (error && error.data && error.data.message) {
-        showSnackbar(error.data.message, "error");
-      }
-    }
-  };
-
-  const handleDeleteColor = async () => {
-    try {
-      await deleteColor({ id: selectedColorId }).unwrap();
-      showSnackbar("Xóa màu sắc thành công!", "success");
-      setOpenDeleteDialog(false);
-      setSelectedColorId(null);
-      refetchColor();
-    } catch (error) {
-      if (error && error.data && error.data.message) {
-        showSnackbar(error.data.message, "error");
-      }
-    }
-  };
-
-  return (
-    <DashboardLayoutWrapper>
-      <Typography variant="h5">Quản lý màu sắc</Typography>
-
-      <Box
-        sx={{ my: 3 }}
-        display={"flex"}
-        justifyContent={"space-between"}
-        alignItems={{
-          xs: "stretch",
-          sm: "center",
-          md: "center",
-        }}
-        flexDirection={{
-          xs: "column",
-          sm: "row",
-          md: "row",
-        }}
-      >
-        <Button
-          variant="outlined"
-          color="primary"
-          onClick={handleRefresh}
-          startIcon={<Refresh />}
-        >
-          Làm mới
-        </Button>
-
-        <Button
-          variant="contained"
-          startIcon={<Add />}
-          onClick={() => {
-            setOpenAddDialog(true);
-            setNewColor({
-              name: "",
-            });
-          }}
-        >
-          Thêm màu sắc
-        </Button>
-      </Box>
-
-      <TableData
-        rows={dataRowColors}
-        totalRows={totalRows}
-        columnsData={columnsColor}
-        loading={isLoadingColor}
-        error={
-          isErrorColor && (
-            <Box mt={2} textAlign="center">
-              <Typography color="error">
-                {errorColor} || "Không tải được dữ liệu."
-              </Typography>
-            </Box>
-          )
-        }
-        paginationModel={paginationModel}
-        onPaginationModelChange={setPaginationModel}
-        pageSizeOptions={[10, 15, 20]}
-      />
-
-      {/* TODO: Dialog add color */}
-      <Dialog fullWidth open={openAddDialog}>
-        <DialogTitle>Thêm màu sắc</DialogTitle>
-        <DialogContent>
-          <TextField
-            label="Tên màu sắc"
-            value={newColor.name}
-            onChange={(e) => setNewColor({ ...newColor, name: e.target.value })}
-            fullWidth
-            sx={{ mt: 2 }}
-            error={submitted && !newColor.name}
-            helperText={
-              submitted && !newColor.name ? "name không được để trống" : ""
-            }
-          />
-        </DialogContent>
-        <DialogActions sx={{ p: 3 }}>
-          <Button
-            color="error"
-            variant="outlined"
-            onClick={() => setOpenAddDialog(false)}
-          >
-            Hủy
-          </Button>
-          <Button onClick={() => handleAddColor(newColor)} variant="contained">
-            Thêm
-          </Button>
-        </DialogActions>
-      </Dialog>
-
-      {/* TODO: Dialog update color */}
-      <Dialog fullWidth open={openUpdateDialog}>
-        <DialogTitle>Cập nhật màu sắc</DialogTitle>
-        <DialogContent>
-          <TextField
-            label="Tên màu sắc"
-            value={newColor.name}
-            onChange={(e) => setNewColor({ ...newColor, name: e.target.value })}
-            fullWidth
-            sx={{ mt: 2 }}
-            error={submitted && !newColor.name}
-            helperText={
-              submitted && !newColor.name ? "name không được để trống" : ""
-            }
-          />
-        </DialogContent>
-        <DialogActions sx={{ p: 3 }}>
-          <Button
-            color="error"
-            variant="outlined"
-            onClick={() => setOpenUpdateDialog(false)}
-          >
-            Hủy
-          </Button>
-          <Button variant="contained" onClick={handleUpdateColor}>
-            Cập nhật
-          </Button>
-        </DialogActions>
-      </Dialog>
-
-      {/* TODO: Dialog delete color */}
-      <Dialog open={openDeleteDialog}>
-        <DialogTitle>Xác nhận xóa ?</DialogTitle>
-        <DialogContent>
-          <Typography>Bạn có chắc chắn muốn xóa màu sắc này không ?</Typography>
-        </DialogContent>
-        <DialogActions sx={{ p: 3 }}>
-          <Button
-            color="error"
-            variant="outlined"
-            onClick={handleCloseDeleteDialog}
-          >
-            Hủy
-          </Button>
-          <Button color="error" variant="contained" onClick={handleDeleteColor}>
-            Xóa
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </DashboardLayoutWrapper>
-  );
-};
-
-export default ColorsManagement;
