+++ src/pages/Admin_page/shared/File_management/index.jsx
@@ -1,0 +1,266 @@
+import { DataGrid } from "@mui/x-data-grid";
+import {
+  Typography,
+  Button,
+  Box,
+  IconButton,
+  Dialog,
+  DialogTitle,
+  DialogContent,
+  DialogContentText,
+  DialogActions,
+} from "@mui/material";
+import DashboardLayoutWrapper from "@/layouts/Dashboard_layout";
+import {
+  useGetAllFilesQuery,
+  useUploadImageMutation,
+  useDeleteImageMutation,
+} from "@/services/api/file";
+import { useState } from "react";
+import { AddPhotoAlternate, Delete, Refresh } from "@mui/icons-material";
+import { PreviewImage } from "@/components/Preview_image";
+import { useSnackbar } from "@/components/Snackbar";
+import TableData from "@/components/Table_data";
+
+
+const FileManagement = () => {
+  const { showSnackbar } = useSnackbar();
+  const [previewImage, setPreviewImage] = useState(null);
+  const [isUploadingImage, setIsUploadingImage] = useState(false);
+  const [paginationModel, setPaginationModel] = useState({
+    page: 0,
+    pageSize: 15,
+  });
+
+  const [openDeleteDialog, setOpenDeleteDialog] = useState(false);
+  const [imageIdToDelete, setImageIdToDelete] = useState(null);
+
+  const {
+    data: dataImage,
+    isLoading: isLoadingImage,
+    isError: isErrorImage,
+    error: errorImage,
+    refetch: refetchImage,
+  } = useGetAllFilesQuery(
+    {
+      page: paginationModel.page + 0,
+      size: paginationModel.pageSize,
+    },
+    {
+      refetchOnMountOrArgChange: true,
+    },
+  );
+
+  const [uploadImage] = useUploadImageMutation();
+  const [deleteImage] = useDeleteImageMutation();
+
+  const dataRowImages = dataImage?.result?.items || [];
+  const totalRows = dataImage?.result?.totalItems || 0;
+
+  const columnsImage = [
+    { field: "id", headerName: "ID", width: 150 },
+    { field: "fileName", headerName: "Tên file", width: 200 },
+    {
+      field: "imageUrl",
+      headerName: "Hình ảnh sản phẩm",
+      width: 200,
+      renderCell: (params) => (
+        <img
+          src={params.row.imageUrl}
+          alt={params.row.fileName}
+          style={{
+            width: 50,
+            height: 50,
+            objectFit: "cover",
+            cursor: "pointer",
+          }}
+          onClick={() => setPreviewImage(params.value)}
+        />
+      ),
+    },
+    {
+      field: "imageUrlShort",
+      headerName: "Mã hình ảnh",
+      width: 200,
+      renderCell: (params) => {
+        if (!params.row || !params.row.imageUrl) {
+          return "N/A";
+        }
+        const urlParts = params.row.imageUrl.split("/");
+        return urlParts[urlParts.length - 1].split(".")[0];
+      },
+    },
+    {
+      field: "actions",
+      headerName: "Hành động",
+      width: 150,
+      renderCell: (params) => (
+        <IconButton
+          color="error"
+          onClick={() => {
+            setImageIdToDelete(params.row.id);
+            setOpenDeleteDialog(true);
+          }}
+        >
+          <Delete />
+        </IconButton>
+      ),
+    },
+  ];
+
+  const handleUploadImage = async (e) => {
+    const file = e.target.files[0]; // Single file only
+    if (!file) return;
+
+    // Validate file type
+    if (!file.type.startsWith("image/")) {
+      showSnackbar("Vui lòng chọn một file hình ảnh!", "warning");
+      return;
+    }
+
+    setIsUploadingImage(true);
+    try {
+      // Pass single file, not array
+      await uploadImage(file).unwrap();
+      showSnackbar("Tải lên hình ảnh thành công!", "success");
+      refetchImage();
+      e.target.value = ""; // Clear input
+    } catch (error) {
+      if (error && error.data && error.data.message) {
+        showSnackbar(error.data.message, "error");
+      }
+    } finally {
+      setIsUploadingImage(false);
+    }
+  };
+
+  const handleDeleteImage = async () => {
+    try {
+      await deleteImage(imageIdToDelete).unwrap();
+      showSnackbar("Xóa hình ảnh thành công!", "success");
+      setOpenDeleteDialog(false);
+      setImageIdToDelete(null);
+      refetchImage();
+    } catch (error) {
+      if (error && error.data && error.data.message) {
+        showSnackbar(error.data.message, "error");
+      }
+    }
+  };
+
+  const handleCloseDeleteDialog = () => {
+    setOpenDeleteDialog(false);
+    setImageIdToDelete(null);
+  };
+
+  const handleRefresh = () => {
+    refetchImage();
+    showSnackbar("Danh sách hình ảnh sản phẩm đã được làm mới!", "info");
+  };
+
+  return (
+    <DashboardLayoutWrapper>
+      <Typography variant="h5">Quản lý hình ảnh sản phẩm</Typography>
+      <Box
+        sx={{
+          mb: 3,
+          mt: 3,
+          gap: {
+            xs: 2,
+            md: 0,
+          },
+        }}
+        display={"flex"}
+        justifyContent={"space-between"}
+        alignItems={{
+          xs: "stretch",
+          sm: "center",
+          md: "center",
+        }}
+        flexDirection={{
+          xs: "column",
+          sm: "row",
+          md: "row",
+        }}
+      >
+        <Button
+          variant="outlined"
+          color="primary"
+          startIcon={<Refresh />}
+          onClick={handleRefresh}
+        >
+          Làm mới
+        </Button>
+        <Button
+          variant="contained"
+          color="primary"
+          component="label"
+          startIcon={<AddPhotoAlternate />}
+          disabled={isUploadingImage}
+        >
+          {isUploadingImage ? "Đang tải..." : "Tải lên hình ảnh"}
+          <input
+            type="file"
+            accept="image/*"
+            hidden
+            onChange={handleUploadImage}
+          />
+        </Button>
+      </Box>
+
+      <TableData
+        rows={dataRowImages}
+        totalRows={totalRows}
+        columnsData={columnsImage}
+        loading={isLoadingImage}
+        error={
+          isErrorImage && (
+            <Box mt={2} textAlign="center">
+              <Typography color="error">
+                {errorImage} || Không tải được dữ liệu.
+              </Typography>
+            </Box>
+          )
+        }
+        paginationModel={paginationModel}
+        onPaginationModelChange={setPaginationModel}
+        pageSizeOptions={[15, 30, 50]}
+      />
+
+      <Dialog open={openDeleteDialog}>
+        <DialogTitle id="delete-dialog-title">
+          Xác nhận xóa hình ảnh ?
+        </DialogTitle>
+        <DialogContent>
+          <DialogContentText id="delete-dialog-description">
+            Bạn có chắc chắn muốn xóa hình ảnh này không ?
+          </DialogContentText>
+        </DialogContent>
+        <DialogActions sx={{ p: 3 }}>
+          <Button
+            onClick={handleCloseDeleteDialog}
+            color="error"
+            variant="outlined"
+          >
+            Hủy
+          </Button>
+          <Button
+            onClick={handleDeleteImage}
+            color="error"
+            variant="contained"
+            autoFocus
+          >
+            Xóa
+          </Button>
+        </DialogActions>
+      </Dialog>
+
+      <PreviewImage
+        previewImage={previewImage}
+        setPreviewImage={setPreviewImage}
+      />
+    </DashboardLayoutWrapper>
+  );
+};
+
+export default FileManagement;
