--- src/services/api/product_image.js
@@ -1,62 +1,0 @@
-import axios from "axios";
-import { baseApi } from "./index";
-import { TAG_KEYS } from "/src/constants/tagKeys.js";
-
-export const productImageApi = baseApi.injectEndpoints({
-  endpoints: (builder) => ({
-    listImages: builder.query({
-      query: ({ page, size, sort }) => ({
-        url: "/v1/admin/files/all",
-        params: { page, size, sort },
-      }),
-      providesTags: [TAG_KEYS.PRODUCT_IMAGE],
-    }),
-
-    uploadImage: builder.mutation({
-      async queryFn(file) {
-        const formData = new FormData();
-
-        formData.append("files", file);
-
-        try {
-          const token = localStorage.getItem("accessToken");
-          const response = await axios.post(
-            `${import.meta.env.VITE_API_URL}/v1/admin/files/upload/images`,
-            formData,
-            {
-              headers: {
-                Authorization: `Bearer ${token}`,
-                // Don't set Content-Type manually, let browser set it with boundary
-              },
-            },
-          );
-
-          return { data: response.data };
-        } catch (error) {
-          return {
-            error: {
-              status: error.response?.status || 500,
-              data: error.response?.data || error.message,
-            },
-          };
-        }
-      },
-      invalidatesTags: [TAG_KEYS.PRODUCT_IMAGE],
-    }),
-
-    deleteImage: builder.mutation({
-      query: (id) => ({
-        url: `/v1/admin/files/delete/${id}`,
-        method: "DELETE",
-      }),
-      invalidatesTags: [TAG_KEYS.PRODUCT_IMAGE],
-    }),
-  }),
-});
-
-export const {
-  useListImagesQuery,
-  useUploadImageMutation,
-  useLazyListImagesQuery,
-  useDeleteImageMutation,
-} = productImageApi;
